requirements.txt
apache-airflow==2.10.2
sqlalchemy==1.4.54
pandas==2.1.4
psycopg2-binary==2.9.9
pyodbc==5.1.0
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite